Have a separate test module
