+++ wasm/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod parser;
+pub mod reader;
+pub mod core;
